% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{add.organisms}
\alias{add.organisms}
\title{add.orgamisms}
\usage{
add.organisms(traj, col = "red")
}
\arguments{
\item{traj}{simulated trajectories from \code{simulateTracjectories}}
}
\description{
Add organic looking polygons to a plot of simulated trajectories
}

