% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{addNoiseBg}
\alias{addNoiseBg}
\title{generate moving "noise" to simulate noise in a video sequence}
\usage{
addNoiseBg(nframe = 30, density = 40, size = 1, duration = 10,
  speed = 10, bg, col.range = c(0, 1))
}
\description{
generates noise (e.g. non-focal species, moving debris, light effects)
param nframe number of frames
param density number of noise particles (over all frames)
param size maximum size of noise in pixels
param duration the maximum number of frames any single generated noise
particle is visible 
param bg previously generated background 
param speed the maximum speed of the moving noise
param col.range color intensity range of noise
}

