% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{calcCost}
\alias{calcCost}
\title{Calculate costs
\code{calcCost} is a function to linking particle i to particle j,
based on particle distance and size.}
\usage{
calcCost(x1, x2, y1, y2, s1, s2, weight = c(1, 1))
}
\arguments{
\item{x1}{x coordinates of particles in frame n.}

\item{x2}{x coordinates of particles in frame n+1.}

\item{y1}{y coordinates of particles in frame n.}

\item{y2}{y coordinates of particles in frame n+1.}

\item{s1}{particle sizes in frame n.}

\item{s2}{particle sizes in frame n+1.}
}
\description{
Calculate costs
\code{calcCost} is a function to linking particle i to particle j,
based on particle distance and size.
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

