% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{confuStats}
\alias{confuStats}
\title{Calculate different statistics for trained neural network.}
\usage{
confuStats(confusion)
}
\arguments{
\item{confusion}{Confusion matrix}
}
\description{
Calculate different statistics for trained neural network.
}

