% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createBackground}
\alias{createBackground}
\title{Background detection}
\usage{
createBackground(m1, m2, m3, d)
}
\arguments{
\item{allFullImages}{Array containing all images (and three color layers)}
}
\value{
Returns array with still background.
}
\description{
\code{createBackground} is a function to create a still background,
exluding movies objects, using loaded image sequences as input, taking
mean pixel values over all frames.
}
\examples{
\dontrun{
stillBack <- createBackground (allFullImages)
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}

