% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{doTrack}
\alias{doTrack}
\title{Track particles
\code{doTrack} is a function to track particles over subsequent frames.
Based on tracking algorithm by Sbalzarini and Koumoutsakos (2005).}
\usage{
doTrack(particleStats = particleStats, L = 50, plot = FALSE,
  images = allFullImagesRGB[[1]], backward = FALSE)
}
\arguments{
\item{particleStats}{blabla}

\item{L}{Cost to link to dummy particle.}

\item{plot}{Default is FALSE. Plot results?}

\item{images}{Original color image (first frame).}

\item{backward}{Default is FALSE. Perform tracking backward?}
}
\value{
Bla
}
\description{
Track particles
\code{doTrack} is a function to track particles over subsequent frames.
Based on tracking algorithm by Sbalzarini and Koumoutsakos (2005).
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}
\concept{
What is the broad searchable concept?
}

