% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{extractMean}
\alias{extractMean}
\title{Find mean particle values}
\usage{
extractMean(ID, colorimages, images, fun = mean)
}
\arguments{
\item{ID}{Particle ID.}

\item{colorimages}{Original color images.}

\item{images}{Array containing identified particles with ID.}

\item{fun}{Default is mean.}
}
\description{
Find mean particle values
}

