% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{extractNeighbors}
\alias{extractNeighbors}
\title{Get R, G and B values for specified coordinates, and its eight neighbor 
pixels.}
\usage{
extractNeighbors(x, y, images)
}
\arguments{
\item{x}{Vector containing x coordinates.}

\item{y}{Vector containing y coordinates.}

\item{Three}{dimensional array.}
}
\description{
Get R, G and B values for specified coordinates, and its eight neighbor 
pixels.
}

