% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{extractRGB}
\alias{extractRGB}
\title{Find values for R, G and B layer for specified coordinates.}
\usage{
extractRGB(x, y, images)
}
\arguments{
\item{x}{Vector containing x coordinates.}

\item{y}{Vector containing y coordinates.}

\item{images}{Three dimensional array.}
}
\description{
Find values for R, G and B layer for specified coordinates.
}

