% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{findPixelRange}
\alias{findPixelRange}
\title{Find pixel range}
\usage{
findPixelRange(colorimages, frame = 1)
}
\arguments{
\item{colorimages}{Array containing original color images.}

\item{frame}{Number specifying which frame to use. Default is frame 1.}
}
\description{
This function can help to find the minimum and maximum particle size
in pixels, to use in \code{identifyParticles}.
}
\examples{
\dontrun{
dir <- "images"
allFullImages <- loadImages (dirPictures=dir,nImages=1:30)
stillBack <- createBackground(allFullImages)
allImages <- subtractBackground(stillBack)
findPixelRange(allFullImages,frame=10)
}
}
\author{
Marjolein Bruijning
}
