% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{generateBackground}
\alias{generateBackground}
\title{generate a background}
\usage{
generateBackground(background = NULL, spots.density = 10, clustering = 0,
  blur = TRUE, blur.coef = 0.025, domain = c("square", "circle"),
  plot = TRUE, sizes = runif(spots.density, 0.1, 1.5))
}
\description{
generates background, or otherwise plots an already generated background
Function that saves simulated trajectories as png files
param background prevously loaded or generated background 
param spots.density density of non-moving spots
param clustering spot clustering parameter
param blur do the spots look burry?(TRUE or FALSE)
param blur.coef spot blurring coefficient
param domain want a specific domain? can be "square" of "circle".
param plot plot the background?  (TRUE or FALSE)
}
\author{
Caspar A. Hallmann, Marjolein Bruijning & Marco D. Visser
}

