% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{getConfMat}
\alias{getConfMat}
\title{Calculate a confusion matrix for trained neural network.}
\usage{
getConfMat(Y, P, thr, stat = "F")
}
\arguments{
\item{Y}{true values}

\item{P}{predictions from NN}

\item{stat}{the statistic of interest}
}
\description{
Calculate a confusion matrix for trained neural network.
}

