% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCoords}
\alias{getCoords}
\title{Get coordinates of all particles}
\usage{
getCoords(m, d)
}
\arguments{
\item{m}{bla}
}
\value{
Returns array with still background.
}
\description{
\code{getCoords} finds all coordinates of values  > 1 in a matrix
}
\author{
Marjolein Bruijning & Marco D. Visser
}

