% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{gMat}
\alias{gMat}
\title{Create random start g matrix
\code{gMat} is a function to create a random G-matrix, that
can be implemented in \code{\link{doTrack}}.}
\usage{
gMat(Phi)
}
\arguments{
\item{Phi}{Cost matrix.}
}
\value{
Random G-matrix
}
\description{
Create random start g matrix
\code{gMat} is a function to create a random G-matrix, that
can be implemented in \code{\link{doTrack}}.
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

