% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{is.TrDm}
\alias{is.TrDm}
\title{is.TrDm}
\usage{
is.TrDm(object)
}
\arguments{
\item{object}{Object to test}
}
\description{
Generic lower-level function to test whether an object
is an TrackDem object.
}

