% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{makeAnimation}
\alias{makeAnimation}
\title{Make animation
\code{makeAnimation} is ...}
\usage{
makeAnimation(particleStats, images = allFullImagesRGB, path = "~",
  incThr = 10, trackRecord = trackRecord)
}
\description{
Make animation
\code{makeAnimation} is ...
}
\author{
Marjolein Bruijning & Marco D. Visser
}

