% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{makeOrg}
\alias{makeOrg}
\title{makeOrg}
\usage{
makeOrg(length = 30, size = 0.05, phi = 0, x = 0, y = 0)
}
\arguments{
\item{length}{number of polygons}

\item{size}{rough size of each organims (radius)}

\item{phi}{shape parameter}

\item{x}{coordinate}

\item{y}{coordinate}
}
\description{
lower-level function to make organic looking polygons
}

