% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{makeTrVaTe}
\alias{makeTrVaTe}
\title{Make a test, validate and training dataset}
\usage{
makeTrVaTe(dat, prop = c(0.8, 0.1, 0.1))
}
\arguments{
\item{dat}{a previously constructed dataset}

\item{prop}{the proportion for each class
c(training, validation,test).}
}
\description{
datasets are randomly assigned to each
category
}

