% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{maxDist}
\alias{maxDist}
\title{Maximum distance}
\usage{
maxDist(x, y)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}
}
\description{
\code{maxDist} is ...
}
\author{
Marjolein Bruining & Marco D. Visser
}

