% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{phiMat}
\alias{phiMat}
\title{Create cost matrix}
\usage{
phiMat(coords1, coords2, sizes1, sizes2, r = 1, L = 50)
}
\arguments{
\item{coords1}{Coordinates of particles in frame n.}

\item{coords2}{Coordinates of particles in frame n+1.}

\item{sizes1}{Sizes of particles in frame n.}

\item{sizes2}{Sizes of particles in frame n+1.}

\item{r}{Default is one; scaling parameter.}

\item{L}{Cost of linking to dummy variable.}
}
\value{
Cost matrix linking particles.
}
\description{
\code{phiMat} is a function to create a cost matrix based on defined
cost function \code{\link{calcCost}}.
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

