% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFunctions.R
\name{print.summaryTrDm}
\alias{print.summaryTrDm}
\title{\code{print} methods for class 'TrDm'.}
\usage{
\method{print}{summaryTrDm}(x, ...)
}
\arguments{
\item{x}{Object of class 'summaryTrDm'.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
\code{print} methods for class 'TrDm'.
}
\author{
Marjolein Bruijning, Caspar A. Hallmann & Marco D. Visser
}
