% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{samplePolar}
\alias{samplePolar}
\title{Polar sampler used in \code{simulateTracjectories}}
\usage{
samplePolar(n = 10, h = 0.05, rho = 0, init = NULL)
}
\arguments{
\item{n}{number of movements}

\item{h}{displacement speed in pixels.}

\item{rho}{correlation parameter for angle of displacement}

\item{init}{initial locations for all particles?}
}
\description{
helper function for \code{simulateTracjectories} to simulates corrected
polar coordinates. See source of \code{simulateTracjectories} for details.
}
\author{
Caspar A. Hallmann, Marjolein Bruijning & Marco D. Visser
}

