% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{saveTrajectory}
\alias{saveTrajectory}
\title{saveTrajectory}
\usage{
saveTrajectory(traj, add.noise = FALSE, name = "trajectory", axes = FALSE,
  background, pars = list(duration = 10, density = 10), width = 480,
  height = NULL, ...)
}
\arguments{
\item{traj}{simulated trajectories from \code{simulateTracjectories} 
#param add.noise if TRUE, moving noise is added}

\item{name}{stem of the filename}

\item{axes}{if TRUE, axes are included}

\item{background}{use if you have a predifined background image of the
format returned by \code{generateBackground}.}

\item{pars}{parameters used to generate moving noise
these include the density (per image) of noise particles and their duration
(in n frames)}
}
\description{
Function that saves simulated trajectories as png files
}
\author{
Caspar A. Hallmann, Marjolein Bruijning & Marco D. Visser
}

