% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{simulateTrajectories}
\alias{simulateTrajectories}
\title{Create a simulated set of movement trajectories}
\usage{
simulateTrajectories(nframes = 20, nIndividuals = 10, h = 0.05, rho = 0,
  domain = c("square", "circle"), correct.boundary = TRUE,
  sizes = runif(nIndividuals) * 0.012 + 0.01, ...)
}
\arguments{
\item{nframes}{number of time frames(steps)}

\item{nIndividuals}{number of individual trajectories}

\item{h}{displacement speed in pixels.}

\item{rho}{correlation parameter for angle of displacement}

\item{domain}{one of "square" or "circle", imposing a [0-1,0-1] rectangle
domain, or a circlular domain of radius 1, respectively.
correct boundary ensure individual trajectories do not cross the domain}

\item{sizes}{Vector of sizes for each simulated particle of length
nIndividuals}
}
\description{
\code{simulateTracjectories} simulates movement trajectories within a 
bounded space, movements are set with speed (h) and may be correlated
in direction (rho). Function simulates movement of particles in a video 
sequence of certain number of frames (nframes) in length.
}
\examples{
\dontrun{
   traj <- simulateTrajectories(nframes=30,h=0.01,rho=0.9)
}
}
\author{
Caspar A. Hallmann, Marjolein Bruijning & Marco D. Visser
}

