% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackSequences.R
\name{substractBackground}
\alias{substractBackground}
\title{Background subtraction}
\usage{
substractBackground(background, images)
}
\arguments{
\item{background}{Array containing still background.}

\item{images}{Array containing all images.}
}
\value{
Returns array with same size as images, subtracted from background.
}
\description{
\code{subtractBackground} is a function to subtract each
image from the created still background.
The objects created through the function contain all moving
pixels.
}
\examples{
\dontrun{

 allImages <- subtractBackground(background=stillBack,allFullImages)
	}
}
\author{
Marjolein Bruining & Marco D. Visser
}
\seealso{
\code{\link{createBackground}}, \code{\link{summary}},
}
\concept{
What is the broad searchable concept?
}

