% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{subtractBackground}
\alias{subtractBackground}
\title{Background subtraction}
\usage{
subtractBackground(bg, colorimages = NULL)
}
\arguments{
\item{bg}{Array containing still background, as returned from
\code{\link{createBackground}}.}

\item{colorimages}{Array containing all frames, obtained by 
\code{\link{loadImages}}. Default is NULL, in case the original images
are used.}
}
\value{
Returns array with same size as images, subtracted from background.
}
\description{
\code{subtractBackground} is a function to subtract each
image from the created still background.
The objects created through the function contain all changing
pixels.
}
\examples{
\dontrun{
allImages <- subtractBackground(stillBack) }
}
\author{
Marjolein Bruijning & Marco D. Visser
}

