% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsFunctions.R
\name{summary.TrDm}
\alias{summary.TrDm}
\title{\code{summary} methods for class 'TrDm'.}
\usage{
\method{summary}{TrDm}(object, incThres = NULL,
  funSize = stats::median, ...)
}
\arguments{
\item{object}{an object of class 'TrDm'.}

\item{incThres}{Minimum length of tracked segments for particles to be
included. By default an automated threshold is calculated. Only for
'tracked' objects.}

\item{funSize}{Statistic to be calculated to obtain particle sizes. By
default \code{median} is used.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
\code{summary} methods for class 'TrDm'.
}
\author{
Marjolein Bruijning, Caspar A. Hallmann & Marco D. Visser
}
