% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{track}
\alias{track}
\title{Track particles
\code{track} is a function to track particles over subsequent frames.
Based on tracking algorithm by Sbalzarini and Koumoutsakos (2005).}
\usage{
track(Phi, g, L = 50)
}
\arguments{
\item{Phi}{Cost matrix.}

\item{g}{Start identify matrix.}

\item{L}{Cost to link to dummy particle.}
}
\value{
List two elements: first contains array with all images,
subset when relevant. Second element contains all original color images as array.
}
\description{
Track particles
\code{track} is a function to track particles over subsequent frames.
Based on tracking algorithm by Sbalzarini and Koumoutsakos (2005).
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

