% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{updateParticles}
\alias{updateParticles}
\title{Apply trained artificial neural network to particleStat object.}
\usage{
updateParticles(nnTrackdemObject, particleStatObject)
}
\arguments{
\item{nnTrackdemObject}{Object of class nnTrackdemObject.}

\item{particleStatObject}{Object of class particleStatObject.}
}
\value{
A list of class particleStatObject.
}
\description{
Apply trained artificial neural network to particleStat object.
}

