% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R, R/trackSequences.R
\name{createBackground}
\alias{createBackground}
\title{Background detection}
\usage{
createBackground(allFullImages, func = NULL)

createBackground(allFullImages, func = NULL)
}
\arguments{
\item{allFullImages}{Array containing all images (and three color layers)}

\item{func}{Function to be performed over each pixel. Default is NULL,
taking mean pixel values.}

\item{allFullImages}{Array containing all images (and three color layers)}

\item{func}{Function to be performed over each pixel. Default is NULL,
taking mean pixel values.}
}
\value{
Returns array with still background.

Returns array with still background.
}
\description{
\code{createBackground} is a function to create a still background,
exluding movies objects, using loaded image sequences as input.

\code{createBackground} is a function to create a still background,
exluding movies objects, using loaded image sequences as input.
}
\examples{
\dontrun{

stillBack <- createBackground (allFullImages)
}
\dontrun{

stillBack <- createBackground (allFullImages)
}
}
\author{
Marjolein Bruining & Marco D. Visser

Marjolein Bruining & Marco D. Visser
}
\seealso{
\code{\link{createBackground}},

\code{\link{createBackground}},
}
\concept{
What is the broad searchable concept?

What is the broad searchable concept?
}

