% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackSequences.R
\name{createBackground}
\alias{createBackground}
\title{Background detection}
\usage{
createBackground(allFullImages, func = mean)
}
\arguments{
\item{allFullImages}{Array containing all images (and three color layers)}

\item{func}{Function to be performed over each pixel. Default is mean.}
}
\value{
Returns array with still background.
}
\description{
\code{createBackground} is a function to create a still background,
exluding movies objects, using loaded image sequences as input.
}
\examples{
\dontrun{

 stillBack <- createBackground (allFullImages,func=mean)
	}
}
\author{
Marjolein Bruining & Marco D. Visser
}
\seealso{
\code{\link{createBackground}}, \code{\link{summary}},
}
\concept{
What is the broad searchable concept?
}

