% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{createBackground}
\alias{createBackground}
\title{Background detection}
\usage{
createBackground(colorimages, method = "mean")
}
\arguments{
\item{colorimages}{Array containing all frames, obtained by 
\code{\link{loadImages}}}
}
\value{
Array with still background.
}
\description{
\code{createBackground} is a function to create a still background,
containing all motionless, by taking
mean pixel values over all frames.
}
\examples{
\dontrun{
stillBack <- createBackground (allFullImages)
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}

