% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{createBackground}
\alias{createBackground}
\title{Background detection}
\usage{
createBackground(colorimages, method = "mean")
}
\arguments{
\item{colorimages}{Array containing all frames, obtained by}

\item{method}{Use method='mean' to calculate the mean value for each pixel and color.
Use 'method='mean2' to deflate dark values (note, this can only be 
used for dark particles on a light background). Use method='filter' to 
replace pixels in which movement has occurred with the mode of neighboring values.
\code{\link{loadImages}}}
}
\value{
Array with still background.
}
\description{
\code{createBackground} is a function to create a still background,
containing all motionless, by taking
mean pixel values over all frames.
}
\examples{
\dontrun{
stillBack <- createBackground (allFullImages)
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}

