% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createBackground}
\alias{createBackground}
\title{Background detection}
\usage{
createBackground(m1, m2, m3, d)
}
\arguments{
\item{m1}{Array containing all frames for first color layer.}

\item{m2}{Array containing all frames for second color layer.}

\item{m3}{Array containing all frames for third color layer.}

\item{d}{Vector containing dimensions of images(number of rows, number
of columns, number of frames).}
}
\value{
Array with still background.
}
\description{
\code{createBackground} is a function to create a still background,
containing all motionless, by taking
mean pixel values over all frames.
}
\examples{
\dontrun{
stillBack <- createBackground (allFullImages)
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}

