% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{createImageSeq}
\alias{createImageSeq}
\title{Create image sequence}
\usage{
createImageSeq(path = "~/Data/", x = 1915, y = 1080, fps = 15,
  nsec = 2)
}
\arguments{
\item{path}{Path path to location of directory containing movies.}

\item{x}{Number of pixels in horizontal direction}

\item{y}{Number of pixels in vertical direction}

\item{fps}{Frames per second, default is 15.}

\item{nsec}{Duration of movie that is exported, default is 2 seconds.}
}
\description{
\code{createImageSeq} is a function to create an image sequence using 
.MTS files as input file. All movies within a directory 'Movies' will
be exported.
}
\author{
Marjolein Bruijning & Marco D. Visser
}

