% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{createTrainingData}
\alias{createTrainingData}
\title{Create trainingdata
\code{createTrainingData} is a function to to create training data that
by manually selecting false and true positives. The created training
data can be implemented in a neural net.}
\usage{
createTrainingData(particleStats, images = allFullImages,
  allImages = allImages, frames = NULL, mId = mId, training = FALSE)
}
\arguments{
\item{particleStats}{A list with particle statistics for each frame.}

\item{images}{An array with the original full color images, in order to plot
on the original images.}

\item{frame}{A number defining the frame that should be used. Default
is NULL; the frame with the maximum number of identified particles is used.}
}
\description{
Create trainingdata
\code{createTrainingData} is a function to to create training data that
by manually selecting false and true positives. The created training
data can be implemented in a neural net.
}
\author{
Marjolein Bruijning & Marco D. Visser
}

