% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{createTrainingData}
\alias{createTrainingData}
\title{Create trainingdata}
\usage{
createTrainingData(particleStats, allFullImagesRGB = allFullImagesRGB,
  allImages = allImages, frames = frames, mId = mId)
}
\arguments{
\item{particleStats}{A list with particle statistics for each frame.}

\item{colorimages}{A list with the original full color images, in order to plot
on the original images.}

\item{frame}{A number defining the frame that should be used. Default
is NULL; the frame with the maximum number of identified particles is used.}
}
\value{
List containing three elements: true positives, false positives,
and the evaluated frame.
}
\description{
\code{createTrainingData} is a function to to create training data that
by manually selecting false and true positives. The created training
data can be implemented in a neural net.
}
\examples{
\dontrun{
manuallySelect(particleStats=trackObject$particleStats,frame=1)
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{manuallySelect}},
}
\concept{
What is the broad searchable concept?
}

