% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{doTrack}
\alias{doTrack}
\title{Track particles}
\usage{
doTrack(particleStats = trackObject$particleStats, L = 50,
  backward = FALSE, sizeMeasure = "n.cell")
}
\arguments{
\item{particleStats}{Output from \code{\link{identifyParticles}}}

\item{L}{Cost for linking to dummy.}

\item{backward}{Default is FALSE.}
}
\value{
Bla
}
\description{
\code{doTrack} is a function link particles using \code{\link{track}}
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

