% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{doTrack}
\alias{doTrack}
\title{Track particles
\code{doTrack} is a function link particles using \code{\link{track}}}
\usage{
doTrack(particles, L = 50, sizeMeasure = "n.cell", weight = weight)
}
\arguments{
\item{particles}{Object with class particleStatistics,
obtained using \code{\link{identifyParticles}}.}

\item{L}{Cost for linking to dummy. Default set at 50.}

\item{sizeMeasure}{Measure for size (area, length, etc.).
Currently not implemented.}

\item{backward}{Reverse frames. Default is FALSE.}
}
\value{
A list of class 'records'. Use 'summary' and 'plot'.
}
\description{
Track particles
\code{doTrack} is a function link particles using \code{\link{track}}
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

