% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{extractInfo}
\alias{extractInfo}
\title{Extract info
\code{extractInfo} creates a dataframe as preparation for 
applying a neural net (\code{\link{runNN}}). For all particles over all
frames, it collects information on color intensities and neighbor pixels.}
\usage{
extractInfo(particles, colorimages = NULL, sbg = NULL, frames = NULL,
  mIdObject = NULL, training = FALSE)
}
\arguments{
\item{mIdObject}{If training=TRUE, provide a list with true and false positives, 
returned from \code{\link{manuallySelect}}, for each frame.}

\item{training}{Logical. Should identified false and true positives 
be combined to this dataframe? Default is FALSE.}

\item{particleStatObject}{A list with particle statistics for each frame,
as obtained by \code{\link{identifyParticles}}.}

\item{images}{An array with the original full color images, in order to plot
on the original images, obtained by \code{\link{loadImages}}.}

\item{mSub}{Images subtracted from background, as obtained by 
\code{\link{subtractBackground}}.}

\item{frame}{A number defining the frame that should be used. Default
is NULL, in that case all frames are used.}
}
\description{
Extract info
\code{extractInfo} creates a dataframe as preparation for 
applying a neural net (\code{\link{runNN}}). For all particles over all
frames, it collects information on color intensities and neighbor pixels.
}
\author{
Marjolein Bruijning & Marco D. Visser
}

