% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{extractMean}
\alias{extractMean}
\title{Find mean and sd particle values}
\usage{
extractMean(ID, colorimages, images, fun = "mean")
}
\arguments{
\item{ID}{Particle ID.}

\item{colorimages}{Original color images.}

\item{images}{Array containing identified particles with ID.}
}
\description{
Find mean and sd particle values
}

