% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{generateBackground}
\alias{generateBackground}
\title{Generate a background}
\usage{
generateBackground(background = NULL, spots.density = 10, clustering = 0,
  blur = TRUE, blur.coef = 0.025, domain = c("square", "circle"),
  plot = FALSE, sizes = NULL)
}
\arguments{
\item{background}{previously loaded or generated background.}

\item{spots.density}{density of non-moving spots.}

\item{clustering}{spot clustering parameter.}

\item{blur}{do the spots look blurry? (\code{TRUE} or \code{FALSE}).}

\item{blur.coef}{spot blurring coefficient.}

\item{domain}{want a specific domain? can be \code{"square"} of \code{"circle"}.}

\item{plot}{Plot the background?  (\code{TRUE} or \code{FALSE}).}

\item{sizes}{Size of noise particles.}
}
\description{
Simulates a background which can contain noise, after generating the trajectories
using \code{\link{simulateTrajectories}}.
}
\author{
Caspar A. Hallmann, Marjolein Bruijning & Marco D. Visser
}

