% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCoords}
\alias{getCoords}
\title{Get coordinates of all particles}
\usage{
getCoords(m, d)
}
\arguments{
\item{m}{Matrix containing values 0's and values > 0.}

\item{d}{Vector containing dimension of matrix.}
}
\value{
Returns matrix with x and y coordinates for all labels > 0.
}
\description{
\code{getCoords} finds all coordinates of values  > 0 in a matrix
}
\author{
Marjolein Bruijning & Marco D. Visser
}

