% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{identifyParticles}
\alias{identifyParticles}
\title{Identify moving particles}
\usage{
identifyParticles(sbg, threshold = -0.1, pixelRange = NULL,
  pthreshold = NULL, select = "dark", colorimages = NULL,
  autoThres = FALSE, perFrame = FALSE, frames = NULL)
}
\arguments{
\item{threshold}{Thresholds for including particles. For a chosen 
threshold for each frame, use pthreshold. Either one value, or a value 
per color layer.}

\item{pixelRange}{Default is NULL. Vector with minimum and maximum particle size, used as a
first filter to identify particles.}

\item{pthreshold}{Default is NULL. If NULL, treshold is used for filter. If
not zero, a threshold based on pthreshold quantile is calculated for each
frame.}

\item{select}{Select dark particles ('dark'), light particles ('light'), or
both ('both'), compared to background. Default is dark.}

\item{autoThres}{TRUE to get an automated threshold for each color layer.
Default is FALSE.}

\item{perFrame}{If autoThres=TRUE, set at TRUE to calculate a threshold for 
each frame seperately. Default is FALSE.}

\item{mSub}{Array containing images containing all moving particles,
as obtained by \code{\link{subtractBackground}}}
}
\value{
This function returns a list with two elements: (1) a list with
particle statistics with identified particles for each frame. (2) An array
containing all binary images.
}
\description{
\code{identifyParticles} is a function to identify particles.
}
\examples{
\dontrun{
  trackObject <- identifyParticles(allImages,threshold=-0.1,pixelRange=c(3,400))
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}

