% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{identifyParticles}
\alias{identifyParticles}
\title{Identify moving particles}
\usage{
identifyParticles(mSub, threshold = -0.1, pixelRange = NULL,
  pthreshold = NULL, select = "negative")
}
\arguments{
\item{mSub}{Array containing images containing all moving particles,
as obtained by \code{\link{subtractBackground}}}

\item{threshold}{Threshold for including particles. For a chosen 
threshold for each frame, use pthreshold.}

\item{pixelRange}{Default is NULL. Vector with minimum and maximum particle size, used as a
first filter to identify particles.}

\item{pthreshold}{Default is NULL. If NULL, treshold is used for filter. If
not zero, a threshold based on pthreshold quantile is calculated for each
frame.}

\item{select}{Use values smaller ('negative') or larger ('positive') then 
threshold. Default is 'negative'.}
}
\value{
This function returns a list with two elements: (1) a list with
particle statistics with identified particles for each frame. (2) An array
containing all binary images.
}
\description{
\code{identifyParticles} is a function to identify particles.
}
\examples{
\dontrun{
  trackObject <- identifyParticles(allImages,threshold=-0.1,pixelRange=c(3,400))
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}

