% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyFunctions.R
\name{identifyParticles}
\alias{identifyParticles}
\title{Identify moving particles}
\usage{
identifyParticles(images, threshold = -0.1, pixelRange = NULL,
  pthreshold = NULL)
}
\arguments{
\item{images}{Array containing images.}

\item{threshold}{Threshold for including particles. For a chosen 
threshold for each frame, use pthreshold.}

\item{pixelRange}{Default is NULL. Vector with minimum and maximum particle size, used as a
first filter to identify particles.}

\item{pthreshold}{Default is NULL. If NULL, treshold is used for filter. If
not zero, a threshold based on pthreshold quantile is calculated for each
frame.}
}
\value{
This function returns a list with two elements: (1) a list with
particle statistics with identified particles for each frame. (2) an array
containing all binary images.
}
\description{
\code{identifyParticles} is a function to identify particles using subtracted
background images.
}
\examples{
\dontrun{

  trackObject <- identifyParticles(allImages,pthreshold=0.001,pixelRange=c(3,400))
}
}
\author{
Marjolein Bruining & Marco D. Visser
}
\concept{
What is the broad searchable concept?
}

