% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{linkTrajec}
\alias{linkTrajec}
\title{Link created track segments}
\usage{
linkTrajec(recordsObject, particleStatObject, L = 50, R = 1,
  incThres = 10)
}
\arguments{
\item{L}{Cost of linking to dummy variable, default is 50.}

\item{R}{Default is one; link to how many subsequent frames?}
}
\value{
Returns a list of class 'records', containing all merged
track segments. See 'summary' and 'plot'.
}
\description{
\code{linkTrajec} is a function to merge track segments, based on 
distance and size of particles, using a record object 
provided by \code{\link{doTrack}}.
recordsObject Object of class records.
particleStatObject Object of class particleStatistics.
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}
}

