% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{linkTrajec}
\alias{linkTrajec}
\title{Create cost matrix}
\usage{
linkTrajec(G, trackRecord = records$trackRecord, label = records$label,
  sizeRecord = records$sizeRecord, L = 50, R = 1,
  particleStats = trackObject$particleStats, incThres = 10)
}
\arguments{
\item{G}{Coordinates of particles in frame n.}

\item{trackRecord}{Coordinates of particles in frame n+1.}

\item{label}{Sizes of particles in frame n.}

\item{sizeRecord}{Sizes of particles in frame n+1.}

\item{L}{Cost of linking to dummy variable.}

\item{R}{Default is one; scaling parameter.}

\item{incThres}{Threshold}

\item{plot}{Default is FALSE. Plot results?}

\item{images}{Default is FALSE. Plot results?}

\item{texxt}{xlab}
}
\value{
Bla
}
\description{
\code{phiMat} is a function to create a cost matrix based on defined
cost function \code{\link{calcCost}}.
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

