% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trackSequences.R
\name{loadImages}
\alias{loadImages}
\title{Informative title}
\usage{
loadImages(direcPictures, filenames = NULL, nImages = 1:50,
  xranges = NULL, yranges = NULL, channel = 1)
}
\arguments{
\item{direcPictures}{DESCRIPTION}

\item{filenames}{DESCRIPTION}

\item{nImages}{DESCRIPTION}

\item{xranges}{DESCRIPTION}

\item{yranges}{DESCRIPTION}

\item{channel}{DESCRIPTION}
}
\value{
what does it return?
}
\description{
\code{loadImages} is a function that (description)
The objects created through the function ...
}
\examples{
\dontrun{

 rnorm(1) # give a example code here




	}
}
\author{
Marjolein Bruining & Marco D. Visser
}
\seealso{
\code{\link{createBackground}}, \code{\link{summary}},
}
\concept{
What is the broad searchable concept?
}

