% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackSequences.R
\name{loadImages}
\alias{loadImages}
\title{Load .png images}
\usage{
loadImages(direcPictures, filenames = NULL, nImages = 1:30,
  xranges = NULL, yranges = NULL)
}
\arguments{
\item{direcPictures}{The path of the folder where the images can be found.}

\item{filenames}{Default is null. If not all files should be loaded, here specify
which files to use.}

\item{nImages}{The total number of images to load.}

\item{xranges}{Default all pixels are loaded; specify to subset the number of columns.}

\item{yranges}{Default all pixels are loaded; specify to subset the number of rows.}
}
\value{
List two elements: first contains array with all images,
subset when relevant. Second element contains all original color images as array.
}
\description{
\code{loadImages} is a function to load png images as a three dimensional array.
The objects created through the function can be used for image analysis.
}
\examples{
\dontrun{

loadAll <- loadImages(direcPictures='~/images1/',filenames=NULL,
nImages=1:30,yranges=1:1080,xranges=1:1915)

}
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{createBackground}}, \code{\link{summary}},
}
\concept{
What is the broad searchable concept?
}

