% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{manuallySelect}
\alias{manuallySelect}
\title{Manual selection of true and false positives.}
\usage{
manuallySelect(particles, colorimages = NULL, frame = NULL)
}
\arguments{
\item{colorimages}{An array with the original full color images, in order to plot
on the original images.}

\item{frame}{A number defining the frame that should be used. Default
is NULL; in that case the frame with the maximum number of identified particles is used.}

\item{particleStats}{A list with particle statistics for each frame,
obtained by \code{\link{identifyParticles}}.}
}
\value{
List containing three elements: true positives, false positives,
and the evaluated frame.
}
\description{
\code{manuallySelect} is a function to to create training data that
by manually selecting false and true positives. The created training
data can be implemented in a neural net.
}
\examples{
\dontrun{
manuallySelect(particleStats=trackObject$particleStats,frame=1)
}
}
\author{
Marjolein Bruijning & Marco D. Visser
}

