% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{runBatch}
\alias{runBatch}
\title{Batch analysis
\code{runBatch} is a function to analyze all image sequences in a specified
directory.}
\usage{
runBatch(path, direcnames = NULL, nImages = 1:30, pixelRange = NULL,
  threshold = -0.1, pthreshold = NULL, select = "negative", nn = NULL,
  incThres = 10, date = TRUE)
}
\arguments{
\item{path}{A character vector of path name.}

\item{direcnames}{Default is NULL. If not all image sequences should be
analyzed, specificy which files to use.}

\item{nn}{Name of artifical neural net if apply it to images. Default
is NULL, resulting in no neural net being applied.}
}
\value{
Dataframe with estimated population size for each image sequence.
}
\description{
Batch analysis
\code{runBatch} is a function to analyze all image sequences in a specified
directory.
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{loadImages}}, \code{\link{createBackground}},
\code{\link{subtractBackground}},\code{\link{identifyParticles}},
\code{\link{trackParticles}}.
}

