% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{saveTrajectory}
\alias{saveTrajectory}
\title{saveTrajectory}
\usage{
saveTrajectory(traj, noise = FALSE, add.noise = FALSE,
  name = "trajectory", axes = FALSE, background, pars = list(duration =
  10, density = 10), ...)
}
\arguments{
\item{traj}{simulated trajectories from \code{simulateTracjectories}}

\item{noise}{if TRUE, background noise is added
#param add.noise if TRUE, moving noise is added}

\item{name}{stem of the filename}

\item{axes}{if TRUE, axes are included}

\item{background}{use if you have a predifined background image of the
format returned by \code{generateBackground}.
#param pars parameters used to generate moving noise
these include the density (per image) of noise particles and their duration
(in n frames)}
}
\description{
Function that saves simulated trajectories as png files
}
\author{
Caspar A. Hallmann, Marjolein Bruijning & Marco D. Visser
}

