% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationFunctions.R
\name{saveTrajectory}
\alias{saveTrajectory}
\title{saveTrajectory}
\usage{
saveTrajectory(traj, add.noise = FALSE, name = "trajectory", axes = FALSE,
  background, pars = list(duration = 10, density = 10), width = 480,
  height = NULL)
}
\arguments{
\item{traj}{Simulated trajectories from \code{simulateTracjectories}.}

\item{add.noise}{Logical. If \code{TRUE}, moving noise is added.}

\item{name}{Stem of the filename.}

\item{axes}{Logical. If \code{TRUE}, axes are included}

\item{background}{Use if you have a predefined background image of the
format returned by \code{generateBackground}.}

\item{pars}{Parameters used to generate moving noise
these include the density (per image) of noise particles and their duration
(in n frames).}

\item{width}{of created png image. By default 480.}

\item{height}{of create png image. If \code{NULL}, \code{width} is used.}
}
\description{
Function that saves simulated trajectories as png files.
}
\author{
Caspar A. Hallmann, Marjolein Bruijning & Marco D. Visser
}

