% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subtractBackground}
\alias{subtractBackground}
\title{Background subtraction}
\usage{
subtractBackground(m1, bg, d)
}
\arguments{
\item{bg}{Array containing still background, as returned from
\code{\link{createBackground}}.}

\item{d}{Vector containing dimensions of images(number of rows, number
of columns, number of frames).}

\item{images}{Array containing all images for one color layer.}
}
\value{
Returns array with same size as images, subtracted from background.
}
\description{
\code{subtractBackground} is a function to subtract each
image from the created still background.
The objects created through the function contain all changing
pixels.
}
\examples{
\dontrun{
## For all color layers
allImages <- sapply(1:3, 
                  function(x) 
                     subtractBackground(allFullImages[,,x,],
                     stillBack[,,x],dim(allFullImages[,,x,])),
                  simplify='array') }
}
\author{
Marjolein Bruijning & Marco D. Visser
}

