% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R, R/trackSequences.R
\name{subtractBackground}
\alias{subtractBackground}
\title{Background subtraction}
\usage{
subtractBackground(background, images)

subtractBackground(background, images)
}
\arguments{
\item{background}{Array containing still background.}

\item{images}{Array containing all images.}

\item{background}{Array containing still background.}

\item{images}{Array containing all images.}
}
\value{
Returns array with same size as images, subtracted from background.

Returns array with same size as images, subtracted from background.
}
\description{
\code{subtractBackground} is a function to subtract each
image from the created still background.
The objects created through the function contain all moving
pixels.

\code{subtractBackground} is a function to subtract each
image from the created still background.
The objects created through the function contain all moving
pixels.
}
\examples{
\dontrun{

 allImages <- subtractBackground(background=stillBack,allFullImages)
}
\dontrun{

 allImages <- subtractBackground(background=stillBack,allFullImages)
}
}
\author{
Marjolein Bruining & Marco D. Visser

Marjolein Bruining & Marco D. Visser
}
\seealso{
\code{\link{createBackground}},

\code{\link{createBackground}},
}
\concept{
What is the broad searchable concept?

What is the broad searchable concept?
}

