% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{testNN}
\alias{testNN}
\title{Train, validate and test artificial
neural networks}
\usage{
testNN(dat, stat = "F", maxH = 5, repetitions = 3, prop = c(8, 1, 1),
  predictors = NULL, ...)
}
\arguments{
\item{dat}{a previously constructed dataset}

\item{maxH}{maximum number of hidden layers to test
note that more layers will require more time to fit.}

\item{repetitions}{the number of repetitions
for the neural network's training.}

\item{prop}{the proportion or ratio for each class
c(training, validation,test).}

\item{predictors}{Optional. A set of custom predictors
for the neural network.}

\item{fits}{statistic. May be "precision", "recall", or
"F" for the harmonic mean of precision and recall.}

\item{\dots}{additional parameters, passed to neuralnet.}
}
\description{
The function fits multiple neural networks to
a dataset that has been randomly assigned to each
of three catergories: train, validate and test.
A final neural net is selected based on a fit statistic
(either precision, recall or the F1-score). All neural networks
are trained to the training dataset. Neural network may vary in
 the number of hidden layers. Classification thresholds are selected
based on the validation data, and then the final neural network
is selected based on the test data.
}
\details{
The neural networks may be selcted based on precision, recall or
a F1-score (default).
In binary classification, precision is the number of correct positive
results divided by the number of all positive predictions. Recall is
the number of correct positive results divided by the number of positive
results that could have been returned if the algoritm was perfect.
A F1 score (F-score/ F-measure) is statistical
measure of accuracy. F1 scores considers both the precision
and the recall. A F1 score may be seen as a weighted average
(harmonic mean) of the precision and recall.
Precision, recall and F1 scores are at best 1 and at worst 0.
}

