% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFunctions.R
\name{trackParticles}
\alias{trackParticles}
\title{Track particles
\code{trackParticles} is a function link particles.}
\usage{
trackParticles(particles, L = 50, R = 2, backward = FALSE,
  sizeMeasure = "n.cell")
}
\arguments{
\item{particles}{Object with class particles,
obtained using \code{\link{identifyParticles}}.}

\item{L}{Cost for linking to dummy. Default set at 50.}

\item{R}{Default is one; link to how many subsequent frames? Default set
at 2.}

\item{backward}{Reverse frames. Default is FALSE.}

\item{sizeMeasure}{Measure for size (area, length, etc.).
Currently not implemented.}
}
\value{
A list of class 'records'. Use 'summary' and 'plot'.
}
\description{
Track particles
\code{trackParticles} is a function link particles.
}
\author{
Marjolein Bruijning & Marco D. Visser
}
\seealso{
\code{\link{doTrack}}, \code{\link{linkTrajec}},
}

