% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnFunctions.R
\name{update.particles}
\alias{update.particles}
\title{Apply trained artificial neural network to particleStat object.}
\usage{
update.particles(particles, neuralnet, ...)
}
\arguments{
\item{nnTrackdemObject}{Object of class nnTrackdemObject.}

\item{particleStatObject}{Object of class particleStatObject.}
}
\value{
A list of class particleStatObject.
}
\description{
Apply trained artificial neural network to particleStat object.
}

